TimeSpan: Replace implementation with core.time : Duration
This reimplements TimeSpan as a small wrapper around `core.time : Duration`.
The intent is to ultimately replace `TimeSpan` with `Duration`.

This will have multiple advantages:
- Less code in Ocean;
- More attention to this code (this hasn't been touched in over a decade);
- Better D2 compatibility (attributes & co);
- Safer and easier API;
- Better compatibility with other type (e.g. `Thread.sleep`);
- Less reliance on platform-specific API;

Like the old utilities in Phobos, the `time` package of Ocean is built around ticks,
with 1 tick equivalent to 1 hnsecs, or 100 nano seconds.

However, Phobos moved away from using ticks directly, as they do not accurately
represent time: for example, some API acception time objects might expect monotonic times,
while other might expect "user time" (leap second and other non monotonic behavior).

In order to adapt Ocean to this distinction, the first step is to reduce the amount
of duplication and special casing, of which removing `TimeSpan` contribute.
This does not actually deprecate `TimeSpan` yet (as many places still rely on it),
but allow to start replacing it.