Logger: Add two extra log levels (Debug, Verbose)
Experience has shown that the log level provided by Ocean's logger
are not always suitable. Notably, they only offer one level for debugging,
with the rest supposed to be reserved for informational messages,
warnings, or more severe events.
However, while debugging an application, it is often necessary to filter
events by different levels of severity or frequency, something not possible
with a single 'Trace' level.