Prometheus: Reduce duplication of test code and imports
The structs definition used in the test code were duplicated three times:
use a single definition. Additionally, a lot of unnecessary imports were
present, most likely because the code was moved from a module to a new one,
and possibly because the code might not have used the Formatter, originally.